requirements.txt
allure-pytest==2.6.5
allure-python-commons==2.6.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
ddt==1.2.1
et-xmlfile==1.0.1
Faker==1.0.7
get==2018.11.19
idna==2.8
jdcal==1.4
more-itertools==7.0.0
openpyxl==2.6.1
pluggy==0.11.0
post==2018.11.20
public==2018.11.20
py==1.8.0
PyMySQL==0.9.3
pypiwin32==223
pytest==4.5.0
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-rerunfailures==7.0
python-dateutil==2.8.0
pywin32==224
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
selenium==3.141.0
six==1.12.0
suds-jurko==0.6
text-unidecode==1.2
urllib3==1.24.1
wcwidth==0.1.7
